Fixed DTR

+++ application/controllers/Schedule.php
@@ -56,8 +56,19 @@ class Schedule extends CI_Controller
 
     public function delete()
     {
-        if (!empty($_POST)) {
-            $this->schedule_model->delete($_POST['schedule_id']);
+        if (!empty($_POST))
+        {
+
+            $res = $this->schedule_model->get_references($_POST['schedule_id']);
+            if (empty($res))
+            {
+                $this->schedule_model->delete($_POST['schedule_id']);
+            }
+            else
+            {
+                $error_msg = "<h6 id='error' style='color: rgba(255, 0,0,0.7); 'hidden>Selected schedule deleted unsuccessfully. The selected schedule is currently used.</h6>";
+                echo $error_msg;
+            }
         }
         redirect(base_url('schedule'));
     }
